+++ src/main/java/dev/blachut/svelte/lang/utils.kt
@@ -1,8 +1,11 @@
 package dev.blachut.svelte.lang
 
+import com.intellij.lang.javascript.psi.JSEmbeddedContent
 import com.intellij.openapi.util.io.FileUtil
 import com.intellij.openapi.util.text.StringUtil
 import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.psi.PsiElement
+import com.intellij.psi.util.PsiTreeUtil
 
 fun isSvelteComponentTag(name: String): Boolean {
     // TODO Support namespaced components
@@ -11,4 +14,9 @@ fun isSvelteComponentTag(name: String): Boolean {
 
 fun getRelativePath(currentFile: VirtualFile, componentFile: VirtualFile): String {
     return FileUtil.getRelativePath(currentFile.parent.path, componentFile.path, '/') ?: ""
+}
+
+fun getJsEmbeddedContent(script: PsiElement?): JSEmbeddedContent? {
+    // JSEmbeddedContent is nested twice, see SvelteJSScriptContentProvider
+    return PsiTreeUtil.getChildOfType(script, JSEmbeddedContent::class.java)?.firstChild as JSEmbeddedContent?
 }
